README and AppVeyor changes.

+++ smart_open/tests/test_smart_open_old.py
@@ -14,6 +14,7 @@ backwards compatibility.
 
 import io
 import logging
+import tempfile
 import os
 import sys
 import hashlib
@@ -27,8 +28,6 @@ import gzip
 
 import smart_open
 from smart_open import smart_open_lib
-
-from .test_smart_open import named_temporary_file
 
 logger = logging.getLogger(__name__)
 
@@ -111,7 +110,7 @@ class SmartOpenHttpTest(unittest.TestCase):
         # TODO: why are these tests writing to temporary files?  We can do the
         # bz2 compression in memory.
         #
-        with named_temporary_file('wb', suffix='.bz2', delete=False) as infile:
+        with tempfile.NamedTemporaryFile('wb', suffix='.bz2', delete=False) as infile:
             test_file = infile.name
 
         with smart_open.smart_open(test_file, 'wb') as outfile:
@@ -672,7 +671,7 @@ class SmartOpenTest(unittest.TestCase):
         text = u'欲しい気持ちが成長しすぎて'
 
         with self.assertRaises(UnicodeEncodeError):
-            with named_temporary_file('wb', delete=True) as infile:
+            with tempfile.NamedTemporaryFile('wb', delete=True) as infile:
                 with smart_open.smart_open(infile.name, 'w', encoding='koi8-r',
                                            errors='strict') as fout:
                     fout.write(text)
@@ -683,7 +682,7 @@ class SmartOpenTest(unittest.TestCase):
         text = u'欲しい気持ちが成長しすぎて'
         expected = u'?' * len(text)
 
-        with named_temporary_file('wb', delete=True) as infile:
+        with tempfile.NamedTemporaryFile('wb', delete=True) as infile:
             with smart_open.smart_open(infile.name, 'w', encoding='koi8-r',
                                        errors='replace') as fout:
                 fout.write(text)
@@ -762,6 +761,10 @@ class WebHdfsWriteTest(unittest.TestCase):
 
 @mock.patch('warnings.warn', mock.Mock())
 class CompressionFormatTest(unittest.TestCase):
+    """
+    Test that compression
+    """
+
     TEXT = 'Hello'
 
     def write_read_assertion(self, test_file):
@@ -771,15 +774,8 @@ class CompressionFormatTest(unittest.TestCase):
         with smart_open.smart_open(test_file, 'rb') as fin:
             self.assertEqual(fin.read().decode('utf8'), self.TEXT)
 
-        try:
+        if os.path.isfile(test_file):
             os.unlink(test_file)
-        except PermissionError:
-            #
-            # Work-around for Windows, see:
-            #
-            # <https://github.com/RaRe-Technologies/smart_open/issues/482>
-            #
-            pass
 
     def test_open_gz(self):
         """Can open gzip?"""
@@ -792,13 +788,13 @@ class CompressionFormatTest(unittest.TestCase):
 
     def test_write_read_gz(self):
         """Can write and read gzip?"""
-        with named_temporary_file('wb', suffix='.gz', delete=False) as infile:
+        with tempfile.NamedTemporaryFile('wb', suffix='.gz', delete=False) as infile:
             test_file_name = infile.name
         self.write_read_assertion(test_file_name)
 
     def test_write_read_bz2(self):
         """Can write and read bz2?"""
-        with named_temporary_file('wb', suffix='.bz2', delete=False) as infile:
+        with tempfile.NamedTemporaryFile('wb', suffix='.bz2', delete=False) as infile:
             test_file_name = infile.name
         self.write_read_assertion(test_file_name)
 
@@ -855,8 +851,9 @@ class MultistreamsBZ2Test(unittest.TestCase):
     )
 
     def create_temp_bz2(self, streams=1):
-        with named_temporary_file('wb', suffix='.bz2', delete=False) as f:
-            f.write(self.DATA * streams)
+        f = tempfile.NamedTemporaryFile('wb', suffix='.bz2', delete=False)
+        f.write(self.DATA * streams)
+        f.close()
         return f.name
 
     def cleanup_temp_bz2(self, test_file):
